+++ ingest2/storage/notify.go
@@ -1,0 +1,6 @@
+package storage
+
+const (
+	ChanNewHistoryLedgerSeq = "new_history_ledger_seq"
+	ChanNewLedgerSeq        = "new_ledgers_seq"
+)